Add plan description and link for a4a sites
